requirements.txt
tinker
tinker-cookbook
datasets
torch
tqdm
pandas
matplotlib
wandb
python-dotenv
